+++ src/components/TodoForm/TodoDeadlineForm.js
@@ -1,0 +1,17 @@
+import React from "react";
+import DatePicker from "react-datepicker";
+import "react-datepicker/dist/react-datepicker.css";
+
+export default function TodoDeadlineForm({ deadline, setDeadline }) {
+
+  return (
+    <div>
+       Deadline
+      <DatePicker
+        selected={deadline}
+        minDate={new Date()}
+        onChange={e => setDeadline(e)}
+      />
+    </div>
+  );
+}
