+++ src/components/TodoForm/helpers/formatDate.js
@@ -1,0 +1,5 @@
+
+export default function formatDate (date) {
+    if (!date) return "";
+    return date.getFullYear() + "-" + (date.getMonth() + 1) + "-" + date.getDate();
+}
